reflect/protoreflect: remove DescriptorProto method
The DescriptorProto method was originally intended to return a raw descriptor.
However, that functionality will be responsibility of the protodesc package
in the near future.

Also document that any callers of Options must import the descriptor package.
This is reasonable since any usage of that method will almost always be of
the form:
	fd.Options().(*descriptorpb.FieldOptions)

Change-Id: I1085f02305faef8dd724e0cb3ad45ac12bd5df0a
Reviewed-on: https://go-review.googlesource.com/c/153497
Reviewed-by: Damien Neil <dneil@google.com>