protoapi: add CompressGZIP helper function
The current generated API relies on GZIP since the Descriptor and
EnumDescriptor methods return the file descriptor as the GZIP-compressed
form of the raw encoded bytes.

It turns out that it was a mistake to compress the descriptor since they
hardly contribute any significant size to the total binary bloat and
instead forces us to have a dependency on gzip in the public API.

The CompressGZIP function added is hand-tailored to be as simple and fast
as possible to reduce runtime cost as well.

Change-Id: I32a62f8ba39a0f2dd12fb31800f270b672facf5f
Reviewed-on: https://go-review.googlesource.com/c/164637
Reviewed-by: Damien Neil <dneil@google.com>