internal/cmd/generate-alias: directly invoke v2 internal_gengo
Now that Cl/167768 removed the dependency on protoc for the generated
test protos, the only remaining dependency is on protoc is for the
import public aliases.

However, we observe that we can avoid protoc since we already have
the file descriptor on hand for the remote targets. Using that,
we can invoke internal_gengo directly and circumvent any need for protoc.

With protoc no longer being a dependency, the integration script
loses its value, so can be deleted.

Change-Id: I522c1e4f2abc47be2817f3bd6b3ff9704b0abb83
Reviewed-on: https://go-review.googlesource.com/c/protobuf/+/167770
Reviewed-by: Herbie Ong <herbie@google.com>