internal/proto: implement SetDefaults with reflection
We cannot start replacing the implementation of v1 directly with reflection
since the scaffolding to avoid cyclic dependencies on descriptor proto
is not yet complete.

However, we can start re-writing functionality in a seperate package as if
they were replacing the version in the proto package.

For the time being, we just copy-paste the relevant tests in the proto
package to ensure the behavior is the same in both implementations.

Change-Id: Ibc9360edb4c8f43bbe60359b848bb2f998c22bf9
Reviewed-on: https://go-review.googlesource.com/c/protobuf/+/167761
Reviewed-by: Herbie Ong <herbie@google.com>