internal/proto: robustify GetProperties for a post-reflection world
Originally, the approach taken was to re-write GetProperties in terms
of protobuf reflection to partially support the use of GetProperties
on generated messages not in the open-struct API.

However, CL/167918 shows that doing so preserves a significant amount of
technical debt in the v1 repository.

In discussing this with @dneil, we decided that we should keep GetProperties
working for open-struct messages, but to instead ensure that GetProperties
does *not* work for non-open-struct messages.

Changes made:
* Document the API more, and place warnings on all exported entry points
deterring away new users.
* Always rename unrelated Go struct fields with an XXX_ prefix since this is
a common pattern that users check for.
* Deliberately fail if the struct does not look like an open-struct API.
* Minor bug fix for groups getting assigned the wrong wire type.

Change-Id: I404bde0327fea9d83387d19fb50feccfa4608e47
Reviewed-on: https://go-review.googlesource.com/c/protobuf/+/168157
Reviewed-by: Herbie Ong <herbie@google.com>