all: move v1 types over to the v2 repository
In CL/168519, we move all exported types in v1 protoapi over
to the v2 repository.

This CL does the other half of the move, where the moved logic is
deleted from protoapi. Since little logic remains in protoapi,
the package is being deleted, and remaining logic moved back into
the v1 proto package itself.

CL/168538 is the corresponding change to alter v1.
There will be a temporary build failure as it is not possible
to submit CL/168519 and CL/168538 atomically.

Change-Id: I3f233b56684357747e35ac6b83f3e0aa427b3152
Reviewed-on: https://go-review.googlesource.com/c/protobuf/+/168538
Reviewed-by: Herbie Ong <herbie@google.com>