proto: unmarshal extensions eagerly
Prior to this CL, Go did not eagerly unmarshal extension fields.
The unmarshaling was done when GetExtension was called.
The prior behavior has always been problematic:
* This is incorrect since we are not validating that extensions
are valid upon the initial unmarshal, which is a property of correctness
that C++ and Java provide.
* Usage is racy. GetExtension is a mutating operation since it can
transform some bytes in the unknown fields to become extension fields.
Thus, concurrent read-only operations with GetExtension is racy.
In v1, we added a lock around the extension map to try and avoid a race,
but this is just a facade. This fixes a race condition in terms of the
Go memory model, but it is still racy since a local operation can still
observe the effects of a remote GetExtension operation.
* The API is clunky since delayed unmarshaling means that GetExtension
needs to deal with possible unmarshal errors.

We make extensions eagerly unmarshaled to avoid all of these problems.

Change-Id: Iafccf2aeaa839c0d3f88ca18bd84af481e9b1662
Reviewed-on: https://go-review.googlesource.com/c/protobuf/+/172399
Reviewed-by: Damien Neil <dneil@google.com>